requirements.txt
RPi.GPIO
RpiMotorLib
pynput